refactor: centralize reveal step counting
