+++ apps/campfire/src/components/Deck/Deck.tsx
@@ -21,7 +21,7 @@ import {
   prefersReducedMotion,
   runAnimation
 } from '@campfire/components/transition'
-import { SlideReveal } from './Slide/SlideReveal'
+import { getRevealMax } from './Slide/utils'
 import type { Transition, SlideTransition } from './Slide/types'
 
 export type ThemeTokens = Record<string, string | number>
@@ -72,34 +72,6 @@ const srOnlyStyle: JSX.CSSProperties = {
   clip: 'rect(0, 0, 0, 0)',
   whiteSpace: 'nowrap',
   border: 0
-}
-
-/**
- * Recursively determines the highest step index contributed by SlideReveal
- * components within a tree.
- *
- * @param children - Slide children to inspect.
- * @returns The maximum step index discovered.
- */
-const getRevealMax = (children: ComponentChildren): number => {
-  let max = 0
-  const walk = (nodes: ComponentChildren): void => {
-    toChildArray(nodes).forEach(node => {
-      if (typeof node === 'object' && node !== null && 'type' in node) {
-        const child = node as VNode<any>
-        if (child.type === SlideReveal) {
-          const at = child.props.at ?? 0
-          const exitAt = child.props.exitAt ?? at
-          max = Math.max(max, at, exitAt)
-        }
-        if (child.props?.children) {
-          walk(child.props.children)
-        }
-      }
-    })
-  }
-  walk(children)
-  return max
 }
 
 /**
